fix: Code refactor for coverage
